ugc/requirements.txt
backoff==2.2.1
fastapi-cache2[redis]==0.2.1
fastapi==0.104.1
gunicorn==21.2.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0
pydantic_settings==2.1.0
pydantic==2.5.2
redis==4.6.0
aiohttp==3.9.3
uvicorn==0.24.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
motor==3.3.2
python-logstash==0.4.8
python-logstash-async==3.0.0